Introduce SDL2
Prior to this change, I had no experience with SDL; but
http://wiki.libsdl.org is very useful indeed.

This change introduces a simple nix derivation and Makefile to build a
binary named mandelbrot from a single and simplistic main.c file.  The
code in main.c is almost entirely lifted from the SDL2 documentation.
More attention is given to logging of error and organising code into
functions.  Besides the main function there are four other functions:
- Setup - contains mostly boilerplate to create a window and render;
- MainLoop - polls for SDL events and calls RenderFrame;
- RenderFrame - calls ClearRenderer and SDL_RenderPresent;
- ClearRenderer - calls SDL_RenderClear.

The main function glues together initialisation of SDL and its logging
system, the Setup and MainLoop functions and teardown of resources.