Log frame rate
Prior to this change, I wanted to measure the frame rate.

This change introduces a static circular buffer to record the clock time
of the last 64 frames.  An estimate of the frame rate is logged every
time the buffer cycles and is simply computed from the first and last
known samples.