Introduce Draw method
Prior to this change, nothing was drawn to the window.

This change introduces the Draw and ComputePixels methods that compute
the pixels for the visualisation of the Mandelbrot set in a given
rectangle and at a given resolution.  A surface is created from the
pixel data using a indexed colour palette from the Sonic Robo Blast 2
project.  A texture is then created from the surface and finally copied
to the window's renderer.

Also the fractal is re-rendered from scratch when the window changes
size.  The frame rate is effected by the resolution, nothing is drawn
while the pixel values are computed.  The resolution is initially set to
a very poor value and incrementally increased with each frame to give
the appearance of a smooth transition from a poor quality picture to a
high quality picture.