Cache computed pixel
Prior to this change, the pixels of the image were recalculated each
frame.  This was ok since the frame wasn't drawn very often.

This change introduces a new struct to caches the pixels along side the
values that were used to compute the pixels.  This shall allow for more
frequent updates of the frame in future.