Improve zooming
Prior to this change, aspect ratio was not preserved when change window
size and zooming was alway focused on the center of the window.

This change introduces separate functions for scaling and translation
and now uses the current mouse location as the focus point of zooming.
When the window size changes a scale factor is chosen to maintain the
width or height of the viewport (which ever is bigger).