Introduce computed palette
Prior to this change, the colour palette was a fixed const taken from
SRB2 and was limited to 256 colours.

This change introduces a computed palette of varying hue from 0 to 360
degrees.  The first and last colours in the palette are reserved for
indicating set membership and drawing orbits.  Although it's still only
limited to 256, the 254 hues can be seemlessly cycled indefinately.