Introduce shell.nix
Prior to this change, nix-build would avoid using cargo to build the
crate and dependencies.  This is great for release but can have problems
of its own.

This change introduces a shell.nix file that will bootstrap the same
environment as default.nix but also include cargo.  This allows for the
more familiar development workflow while still having a high degree of
reproducibility.