Increase detail
Prior to this change, the level of detail detected in the fractal was
limited to the number of colours (254) used to represent the number of
steps that an orbit took to reach the divergance limit (DIV_LIMIT).
Failing to reach this limit within the 254 steps means that the value is
declared convergent, which could be premature.  Not computing more steps
to discover more divergant values leads to a lack of detail at a higher
scale factor.

This change increases the number of colours by cycling through the
available 254, ten times, determined by FIXED_THRESHOLD.  (The number of
cycles is arbritary and improves the amount of detail seen at high scale
factors, at the current level the loss become apparent at
tx=0.0000007+0i.  Also, colour index=0 is now considered to be
unspecified and index=255 is considered convergent.

Note that a loss of precision (due to decimal approximation) becomes
apparent after a scale of 0.00000000000000004.