Refactor orbit compute logic into pure function
Prior to this change, the compute and compute_orbit functions share some
common logic, the former tailored to count the number of steps and the
latter to construct the intermediatary values of each step.  This is the
core logic of the whole program and deserves to be pure (also it will be
easier to pararellise the computation).

This change removes compute_orbit as a method on image and promotes it
to a top level function.  It now optionally takes a Vec to record the
orbit and will only return the number of steps to reach DIV_LIMIT if it
actually surpasses the limit.

Note that there is an allocation per orbit of a Complex64 struct
introduced for the compute method.  But it's possible that rustc
can/will optimise it away(?)