Parallelise compute
Prior to this change, compute would call compute_orbit for each pixel in
the image in serial.

This change will create a set of threads in compute and a channel for
each thread.  Then it feeds pixel index and value pairs for each pixel
and distributes the pairs evenly across the threads.  Each thread places
the result onto another which is consumed by the main thread.

Note that distributing the pixels evenly is not the same as distributing
the workload evenly.  Also the call to compute is still on the main
thread and locks the application from accepting user input.