Introduce transformation between disparate types
Prior to this change, the components of the the Source and Projected
types had to be the same, i.e. f64 -> f64.  This doesn't align with the
mapping of points on the complex plane to pixels on the sceen.

This change introduces TryProjector trait which allows lossy
transformations.  This trait is used to map Complex<Source<f64>> type to
Complex<Projected<i32>> type.

Signed-off-by: Corin Lawson <corin@responsight.com>