Ensure the first line of trace is always drawn
Prior to this change, if the trace of the orbit wasn't able to be
represented in i32 then the trace couldn't be drawn.

This change bifucates the first line until it can be represent with i32
values and draws it, thus guaranteeing that at least one line will be
drawn to indicate that the trace is still active.